requirements.txt
PyMuPDF==1.23.0
Pillow==10.0.0
json5==0.9.0
pytest==7.4.0